Cleanup on-demand tiling tests Add type-related test (issue #122)
